backend/requirements.txt
# Django & REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0

# Third-party packages
django-cors-headers>=4.3,<5.0
python-dotenv>=1.0,<2.0
djangorestframework-simplejwt>=5.3,<6.0

# Database
psycopg2-binary>=2.9,<3.0

# Production
gunicorn>=21.0,<22.0
whitenoise>=6.6,<7.0
